refactor: move tsc transformer to a separate file
